pip_requirements.txt
certifi==2022.6.15
cloudpickle==2.0.0
cycler==0.11.0
cytoolz==0.11.0
dask==2022.5.0
fonttools==4.25.0
fsspec==2022.3.0
imagecodecs==2021.8.26
imageio==2.9.0
kiwisolver==1.4.2
locket==1.0.0
matplotlib==3.4.3
mkl-service==2.4.0
munkres==1.1.4
networkx==2.7.1
numpy==1.16.6
packaging==21.3
partd==1.2.0
Pillow==9.0.1
pip==21.2.4
pyparsing==3.0.4
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.19.2
scipy==1.7.3
setuptools==61.2.0
sip==4.19.13
six==1.16.0
tifffile==2021.7.2
toolz==0.11.2
tornado==6.1
wheel==0.37.1
